Cambie las posiciones a double.

+++ trunk/src/titiritero/Posicionable.java
@@ -6,7 +6,7 @@ package titiritero;
  */
 public interface Posicionable {
 	
-	public int getX();
-	public int getY();
+	public double getX();
+	public double getY();
 
 }
